Refactor search code into seperate files from engine. Place constants in constants file.
