+++ src/lib.rs
@@ -2,3 +2,5 @@ pub mod ordering;
 pub mod engine;
 pub mod evaluation;
 pub(crate) mod consts;
+pub(crate) mod quiescent;
+pub(crate) mod search;
